+++ VeniVidiDaVinci/src/GameScreenLevelEditor.h
@@ -30,6 +30,6 @@ public:
 	~GameScreenLevelEditor();
 
 	void Render();
-	void Update(double deltaTime, SDL_Event event);
+	void Update(float deltaTime, SDL_Event event);
 };
